Add plotting package with OCF brand theme (closes #95)
Introduces packages/plotting with ocf_theme (palette constants + Altair
theme registration) and forecast_chart (build_forecast_chart, moved from
plot_jobs.py). All charts now use the OCF colour scheme; hardcoded
"steelblue" / "lightgray" / "teal" replaced with OCF constants.
