+++ packages/dashboard/pyproject.toml
@@ -8,6 +8,7 @@ dependencies = [
     "altair>=6.0.0",
     "lonboard>=0.13.0",
     "marimo>=0.19.7",
+    "plotting",
     "polars>=1.37.1",
     "pyarrow>=23.0.0",
     "geoarrow-pyarrow>=0.2.0",
