+++ packages/plotting/pyproject.toml
@@ -1,0 +1,13 @@
+[project]
+name = "plotting"
+version = "0.1.0"
+description = "OCF brand theme and chart-building utilities"
+requires-python = ">=3.14"
+dependencies = [
+    "altair>=6.0.0",
+    "polars>=1.0.0",
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
